backend/requirements.txt
django
djangorestframework
django-cors-headers
django-channels
psycopg2-binary
djangorestframework-simplejwt
Pillow
Pillow